Start of some refactorings but looking good.
PatuiStepData was useless as time was in PatuiEvent so PatuiStepDataFlavour has become just PatuiStepData (maybe should be PatuiData even?).

Started to move eval out of assertion step, next I think is to change it so that it's ignorant of results passed in and put the logic around step lengths in the assertion step.